Extract Python and Dask `Executor` classes from `Workflow` (#1609)
* Extract `MerlinPythonExecutor` from `nvt.Workflow`

* Extract `MerlinDaskExecutor` from `nvt.Workflow`

* Clean up `MerlinDaskExecutor`

* Clean up `MerlinPythonExecutor`

* Move `_clear_worker_cache` to `MerlinDaskExecutor`

* Move `ensure_optimize_dataframe_graph` into `MerlinDaskExecutor`

* Clarify `Nodes` vs `Operators` in `Workflow.fit()`

* Extract the Dask-specific part of `Workflow.fit` to `MerlinDaskExecutor`

* Move Dask client into `MerlinDaskExecutor`

* Inline `_get_stat_op_nodes` to improve clarity

* Clean up `MerlinDaskExecutor.fit()`
